fix: make epoch and rlnIdentifier part of RLNFullProof
`externalNullifier = h(epoch, rlnIdentifier)`
Even though we only need externalNullifier and the other public inputs
to verify a proof, we still need the original epoch and rlnIdentifier
to determine if it's valid for an app. And thus this commit makes
RLNFullProof contain the proof with public inputs, additionally,
`epoch` and `rlnIdentifier`.